requirements.txt
# MetaTrader 5 Python Integration
MetaTrader5>=5.0.45

# Data manipulation (if needed for backtesting)
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0

# Required for production
pyyaml>=6.0
python-dateutil>=2.8.2